\name{post_to_data_table}
\alias{post_to_data_table}
\title{
Convert extracted samples
}
\description{
Takes samples from stan model (list) and converts to \code{\link{data.table}}.
}
\usage{
post_to_data_table(pp)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{pp}{posterior samples extracted from stan with \code{\link{rstan::extract}}}
}
\details{

}
\value{
samples formated as \code{\link{data.table}}
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Mark
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}

